build: update build process to only bundle .js and .d.ts files
This PR is meant to stop packaging full TypeScript files in the
library. This breaks testing capabilities in Angular, and is
overall unnecessary. Instead, we transpile all files using tsc
into a separate directory, and only publish that directory, with
the root files (index.js and index.d.ts) specified in package.json.

There are no breaking changes in this PR, but if you are building
from source, please make sure to update your dev dependencies.

* add outDir to tsconfig for bundling
* upgrade to angular v5 and use new compiler instead of just tsc
* clean up package.jsom npm scripts
* add exports to index.ts to allow for root import
* move inline builder to separate directory
* add dev dependencies for TypeScript building
* add files and directories to npmignore to allow for root packaging
* move protractor config to root level
* change webpack to use ts-loader

* Fixes #49

BREAKING CHANGE:

Due to the change in how folders are exported, all imports from
`preboot` must come as a root import, e.g.

`import {ServerPrebootModule} from 'preboot';` instead of
`import {ServerPrebootModule} from 'preboot/server';`
