+++ package.json
@@ -1,15 +1,17 @@
 {
   "name": "preboot",
-  "version": "5.1.7",
+  "version": "6.0.0",
   "description": "Record server view events and play back to Angular client view",
+  "main": "lib/index.js",
+  "types": "lib/index.d.ts",
   "scripts": {
-    "inline": "rm -rf dist && mkdir dist && node src/build/inline.preboot.builder || true",
+    "inline": "ts-node build/inline.preboot.builder || true",
     "rmrf": "rm -rf node_modules && rm -rf package-lock.json || true",
-    "build": "rm -rf node_modules && npm i && ./node_modules/.bin/webdriver-manager update && tsc && npm run inline || true",
-    "test": "./node_modules/.bin/jasmine JASMINE_CONFIG_PATH=test/utils/jasmine.json || true",
-    "start": "node test/e2e/e2e.server start || true",
-    "e2e:build": "./node_modules/.bin/webdriver-manager update && tsc && node test/e2e/e2e.server.gen.js && node_modules/.bin/webpack --config test/e2e/webpack.config.js || true",
-    "e2e:test": "./node_modules/.bin/protractor test/e2e/protractor.conf || true",
+    "build": "rm -rf node_modules && npm i && rm -rf lib && webdriver-manager update && mkdir lib && ngc -p tsconfig.json && npm run inline || true",
+    "test": "jasmine JASMINE_CONFIG_PATH=test/utils/jasmine.json || true",
+    "start": "ts-node test/e2e/e2e.server start || true",
+    "e2e:build": "webdriver-manager update && ngc -p tsconfig.json && ts-node test/e2e/e2e.server.gen.ts && webpack --config test/e2e/webpack.config.js || true",
+    "e2e:test": "protractor ./protractor.conf || true",
     "e2e": "npm run e2e:build && npm run e2e:test || true"
   },
   "author": "Jeff Whelpley",
@@ -20,14 +22,15 @@
     "Jeff Whelpley <jeff@gethuman.com>"
   ],
   "devDependencies": {
-    "@angular/animations": "^4.4.6",
-    "@angular/common": "^4.4.6",
-    "@angular/compiler": "^4.4.6",
-    "@angular/core": "^4.4.6",
-    "@angular/http": "^4.4.6",
-    "@angular/platform-browser": "^4.4.6",
-    "@angular/platform-browser-dynamic": "^4.4.6",
-    "@angular/platform-server": "^4.4.6",
+    "@angular/animations": "^5.0.0-rc.9",
+    "@angular/common": "^5.0.0-rc.9",
+    "@angular/compiler": "^5.0.0-rc.9",
+    "@angular/compiler-cli": "^5.0.0-rc.9",
+    "@angular/core": "^5.0.0-rc.9",
+    "@angular/http": "^5.0.0-rc.9",
+    "@angular/platform-browser": "^5.0.0-rc.9",
+    "@angular/platform-browser-dynamic": "^5.0.0-rc.9",
+    "@angular/platform-server": "^5.0.0-rc.9",
     "@types/express": "^4.0.38",
     "@types/jasmine": "^2.6.1",
     "@types/node": "^8.0.47",
@@ -40,10 +43,13 @@
     "protractor": "^5.2.0",
     "reflect-metadata": "^0.1.10",
     "rxjs": "^5.5.1",
+    "ts-loader": "^3.1.1",
+    "ts-node": "^3.3.0",
+    "typescript": "^2.5.3",
     "uglify-js": "^3.1.5",
     "webdriver-manager": "^12.0.6",
     "webpack": "^3.8.1",
-    "zone.js": "^0.8.14"
+    "zone.js": "^0.8.18"
   },
   "dependencies": {},
   "peerDependencies": {},
