+++ protractor.conf.js
@@ -6,7 +6,7 @@ const { SpecReporter } = require('jasmine-spec-reporter');
 exports.config = {
   allScriptsTimeout: 11000,
   specs: [
-    '../e2e/**/*.spec.js'
+    './test/e2e/**/*.spec.ts'
   ],
   capabilities: {
     'browserName': 'chrome'
@@ -20,6 +20,9 @@ exports.config = {
     print: function () { }
   },
   onPrepare() {
+    require('ts-node').register({
+        project: 'test/e2e/tsconfig.e2e.json'
+    });
     jasmine.getEnv().addReporter(new SpecReporter({ spec: { displayStacktrace: true } }));
   }
 };
