--- record.ts
@@ -1,1 +1,0 @@
-export { getInlinePrebootCode } from './src/server/inline.preboot.code';
