--- replay.ts
@@ -1,1 +1,0 @@
-export { EventReplayer } from './src/browser/event.replayer';
