+++ src/index.ts
@@ -1,3 +1,5 @@
 export * from './browser';
 export * from './common';
 export * from './server';
+export * from './record';
+export * from './replay';
