+++ src/record.ts
@@ -1,0 +1,1 @@
+export { getInlinePrebootCode } from './server/inline.preboot.code';
