+++ src/replay.ts
@@ -1,0 +1,1 @@
+export { EventReplayer } from './browser/event.replayer';
