+++ test/e2e/e2e.browser.module.ts
@@ -4,7 +4,7 @@ import 'zone.js/dist/long-stack-trace-zone.js';
 
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
-import { BrowserPrebootModule } from '../../src/browser';
+import { BrowserPrebootModule } from '../../lib/browser';
 import { AppComponent, CURRENT_PLATFORM } from './e2e.app';
 
 @NgModule({
