+++ test/e2e/e2e.server.module.ts
@@ -1,7 +1,7 @@
 import { NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { ServerModule } from '@angular/platform-server';
-import { ServerPrebootModule } from '../../src/server';
+import { ServerPrebootModule } from '../../lib/server';
 import { AppComponent, CURRENT_PLATFORM } from './e2e.app';
 
 @NgModule({
