+++ test/e2e/webpack.config.js
@@ -3,14 +3,19 @@ const path = require('path');
 const rootDir = process.cwd();
 
 module.exports = {
-  entry: path.resolve(rootDir, 'test/e2e/e2e.browser.js'),
+  entry: path.resolve(rootDir, 'test/e2e/e2e.browser.ts'),
+  target: 'node',
+  resolve: { extensions: ['.js', '.ts'] },
   output: {
     path: path.resolve(rootDir, 'test/e2e/dist'),
-    filename: 'e2e.browser.js'
+    filename: '[name].js'
   },
 
   // todo: remove this if causing any problems
   module: {
+    rules: [
+      { test: /\.ts$/, loader: 'ts-loader' },
+    ],
     exprContextCritical: false
   },
   plugins: [
